chore: update cli
