+++ packages/mona-cli/package.json
@@ -21,6 +21,7 @@
   "license": "MIT",
   "dependencies": {
     "@bytedance/mona-service": "0.2.0-alpha.9",
+    "@bytedance/mona-shared": "0.2.0-alpha.9",
     "chalk": "^4.1.2",
     "command-line-usage": "^6.1.1",
     "compare-version": "^0.1.2",
