+++ packages/mona-cli/src/index.ts
@@ -1,17 +1,9 @@
-import path from 'path';
-import { GlobalService } from '@bytedance/mona-service'
+import { Service } from '@bytedance/mona-service'
+import buildInPlugins from './buildInPlugins';
 
-const pathToPlugin = (pathname: string) => require(pathname);
-
-const buildInPlugins = [
-  './commands/init',
-  './commands/login',
-  './commands/logout',
-  './commands/update'
-].map(name => pathToPlugin(path.join('@bytedance/mona-cli/dist', name)))
 
 function mona() {
-  const service = new GlobalService(buildInPlugins);
+  const service = new Service(buildInPlugins);
 
   service.install();
